Unify into flat components, move lazy modules into app root
