+++ projects/playground/src/app/components/lazy-module.component.ts
@@ -2,8 +2,8 @@ import { Component, OnInit } from '@angular/core';
 
 @Component({
   selector: 'app-lazy-module',
-  templateUrl: './lazy-module.component.html',
-  styleUrls: ['./lazy-module.component.css'],
+  template: '<p data-cy="lazy-module">lazy-module works!</p>',
+  styles: [],
 })
 export class LazyModuleComponent implements OnInit {
   constructor() {}
