+++ projects/playground/src/app/lazy-module.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { LazyModuleRoutingModule } from './lazy-module-routing.module';
-import { LazyModuleComponent } from './lazy-module.component';
+import { LazyModuleComponent } from './components/lazy-module.component';
 
 @NgModule({
   declarations: [LazyModuleComponent],
