--- projects/playground/src/app/lazy-module/lazy-module.component.html
@@ -1,1 +1,0 @@
-<p data-cy="lazy-module">lazy-module works!</p>
