--- projects/playground/src/app/protected/protected.component.html
@@ -1,1 +1,0 @@
-<p data-cy="protected">This route is protected!</p>
