--- projects/playground/src/app/unprotected/unprotected.component.html
@@ -1,1 +1,0 @@
-<p data-cy="unprotected">This route is unprotected!</p>
